cli: move codeowners lib into "new" module
